Cleaned up project structure
* Created subfolders for hardware (sensors, switchs, etc.) and display (templates, bitmaps)
* Renamed some header files and classes in accordance with C/C++ naming conventions
* Clean up remaining formatting issues
