+++ src/PeriodicTrigger.h
@@ -1,8 +1,8 @@
 /**
  * @file PeriodicTrigger.h
  *
- * @brief TODO
- *
+ * @brief Programmable timer interval
+ * 
  */
 
 #pragma once
@@ -11,8 +11,20 @@ class PeriodicTrigger {
  public:
     PeriodicTrigger(unsigned long millisec);
 
+    /**
+     * @brief Check if the interval period has elapsed
+     * @return Returns true until the timer is reset
+    */
     bool check();
+
+    /**
+     * @brief Reset timer interval to currently configured value
+    */
     void reset();
+
+    /**
+     * @brief Reset timer with new timer interval
+    */
     void reset(unsigned long millisec);
 
  private:
