+++ src/isr.h
@@ -1,13 +1,17 @@
 /**
- * @file ISR.cpp
+ * @file isr.h
  *
- * @brief TODO
+ * @brief Timer - ISR for PID calculation and heater relay output
  *
  */
 
+#pragma once
+
 #include <Arduino.h>
-#include "ISR.h"
-#include "pinmapping.h"
+#include "hardware/pinmapping.h"
+
+extern double pidOutput;
+extern hw_timer_t *timer;
 
 unsigned int isrCounter = 0;  // counter for ISR
 unsigned long windowStartTime;
@@ -26,32 +30,28 @@ void IRAM_ATTR onTimer(){
 
     isrCounter += 10; // += 10 because one tick = 10ms
 
-    //set PID output as relais commands
+    // set PID output as relay commands
     if (isrCounter >= windowSize) {
         isrCounter = 0;
     }
 }
 
-
 /**
  * @brief Initialize hardware timers
  */
 void initTimer1(void) {
-    timer = timerBegin(0, 80, true); //m
-    timerAttachInterrupt(timer, &onTimer, true);//m
-    timerAlarmWrite(timer, 10000, true);//m
+    timer = timerBegin(0, 80, true);
+    timerAttachInterrupt(timer, &onTimer, true);
+    timerAlarmWrite(timer, 10000, true);
 }
-
 
 void enableTimer1(void) {
     timerAlarmEnable(timer);
 }
 
-
 void disableTimer1(void) {
     timerAlarmDisable(timer);
 }
-
 
 bool isTimer1Enabled(void) {
     return timerAlarmEnabled(timer);
